Fix: cast input to float32 before update_grid (bfloat16 not supported)
torch.linalg.lstsq inside efficient_kan's update_grid only supports
float32 and float64. With bfloat16 training this caused a crash at
step 10000 (the first grid update, configured at update_grid_every=10000).

Fix: cast x to float32 in KANEncoder.update_grid before passing to
KANLinear.update_grid. The result is copied back into spline_weight
(bfloat16) via .data.copy_(), which does the dtype conversion automatically.

Training confirmed working: at step 10000 before crash, recon=2.13 (down
from 751 at step 0), active=12/4096 features, val_loss=2.0578 at step 5000.

+++ kan_clt/kan_encoder.py
@@ -186,4 +186,5 @@ class KANEncoder(nn.Module):
         """
         if x.dim() > 2:
             x = x.reshape(-1, self.d_model)
-        self.kan_linear.update_grid(x)
+        # update_grid uses torch.linalg.lstsq which doesn't support bfloat16
+        self.kan_linear.update_grid(x.float())
